\name{Abiotic_gates}
\docType{data}
\encoding{UTF-8}
\alias{Abiotic_gates}
\title{Dataset with abiotic parameters and gate information.}
\description{Example dataset of abiotic parameters and gate information for every sample based on \emph{CyBar} analysis (see reference).}
\usage{data(Abiotic_gates)}
\format{Data frame with 20 observations of 10 variables.}
\references{\emph{Christin Koch, Ingo Fetzer, Thomas Schmidt, Hauke Harms and Susann Müller. Monitoring functions in managed microbial systems by cytometric bar coding. Environmental Science & Technology, 2013.}
}
\keyword{datasets}