\name{abiotic_incol}
\docType{data}
\alias{abiotic_incol}
\title{Dataset with abiotic data}
\description{Example dataset containing abiotic parameters used for the NMDS plot of the second downloadable dataset (see manual).}
\usage{data(abiotic_incol)}
\format{Data frame with 17 observations of 6 variables.}
\value{Abiotic data of the second downloadable dataset.}
\keyword{datasets}