\name{flowCHIC-package}
\encoding{UTF-8}
\alias{flowCHIC}
\docType{package}
\title{
Analyze flow cytometric data using histogram information
}
\description{
A package to analyze flow cytometric data of complex microbial communities based on histogram images.
}
\details{
\tabular{ll}{
Package: \tab flowCHIC\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2014-11-26\cr
License: \tab GPL-2\cr
}
abiotic_incol
fcs_to_img
flowCHIC
img_sub
calculate_overlaps_xor
plot_nmds
Results_xor_mix
Results_xor_incol
Results_overlaps_mix
Results_overlaps_incol
}
\author{Joachim Schumann \email{joachim.schumann@ufz.de}, Christin Koch \email{christin.koch@ufz.de}, Ingo Fetzer \email{info.fetzer@stockholmresilience.su.se}, Susann Müller \email{susann.mueller@ufz.de}
}
\references{\emph{Christin Koch, Ingo Fetzer, Hauke Harms, and Susann Müller. CHIC - An Automated
Approach for the Detection of Dynamic Variations in Complex Microbial Communities.
Cytometry Part A, 2013.}}

\seealso{\code{\link[flowCore:flowCore-package]{flowCore}},\code{\link[flowViz:flowViz-package]{flowViz}},\code{\link[flowCyBar:flowCyBar-package]{flowCyBar}}}

\keyword{package}