\name{Results_overlaps}
\docType{data}
\alias{Results_overlaps}
\title{Dataset with labels and sums of informative pixels within the overlap area of the FCS files included to the package}
\description{Example dataset containing labels and areas of each image pair of the FCS files included to the package.}
\usage{data(Results_overlaps)}
\format{Data frame with 3 observations of 2 variables.}
\value{Results of the method \emph{calculate_overlaps_xor()} on the FCS files included to the package.}
\keyword{datasets}