\name{Results_overlaps_incol}
\docType{data}
\alias{Results_overlaps_incol}
\title{Dataset with labels and sums of informative pixels within the overlap area of the second downloadable dataset}
\description{Example dataset containing labels and areas of each image pair of the second downloadable dataset (see manual).}
\usage{data(Results_overlaps_incol)}
\format{Data frame with 136 observations of 2 variables.}
\value{Results of the method \emph{calculate_overlaps_xor()} on the subsets of the second downloadable dataset.}
\keyword{datasets}