\name{Results_overlaps_mix}
\docType{data}
\alias{Results_overlaps_mix}
\title{Dataset with labels and sums of informative pixels within the overlap area of the first downloadable dataset}
\description{Example dataset containing labels and areas of each image pair of the first downloadable dataset (see manual).}
\usage{data(Results_overlaps_mix)}
\format{Data frame with 4005 observations of 2 variables.}
\value{Results of the method \emph{calculate_overlaps_xor()} on the subsets of the first downloadable dataset.}
\keyword{datasets}