\name{Results_xor}
\docType{data}
\alias{Results_xor}
\title{Dataset with labels and intensity values of the FCS files included to the package}
\description{Example dataset containing labels and intensity values of each image pair of the FCS files included to the package.}
\usage{data(Results_xor)}
\format{Data frame with 3 observations of 2 variables.}
\value{Results of the method \emph{calculate_overlaps_xor()} on the FCS files included to the package.}
\keyword{datasets}