\name{Results_xor_incol}
\docType{data}
\alias{Results_xor_incol}
\title{Dataset with labels and intensity values of the second downloadable dataset}
\description{Example dataset containing labels and intensity values of each image pair of the second downloadable dataset (see manual).}
\usage{data(Results_xor_incol)}
\format{Data frame with 136 observations of 2 variables.}
\value{Results of the method \emph{calculate_overlaps_xor()} on the subsets of the second downloadable dataset.}
\keyword{datasets}