\name{Results_xor_mix}
\docType{data}
\alias{Results_xor_mix}
\title{Dataset with labels and intensity values of the first downloadable dataset}
\description{Example dataset containing labels and intensity values of each image pair of the first downloadable dataset (see manual).}
\usage{data(Results_xor_mix)}
\format{Data frame with 4005 observations of 2 variables.}
\value{Results of the method \emph{calculate_overlaps_xor()} on the subsets of the first downloadable dataset.}
\keyword{datasets}