\name{flowCHIC-package}
\encoding{UTF-8}
\alias{flowCHIC}
\docType{package}
\title{
Analyze flow cytometric data using histogram information
}
\description{
A package to analyze flow cytometric data of complex microbial communities based on histogram images
}
\details{
\tabular{ll}{
Package: \tab flowCHIC\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2014-02-13\cr
License: \tab GPL-2\cr
}
Abiotic_data
Abiotic_gates
flowCHIC
chic
Groups
Results_overlaps
Results_xor

}
\author{Joachim Schumann \email{joachim.schumann@ufz.de}, Christin Koch \email{christin.koch@ufz.de}, Ingo Fetzer \email{info.fetzer@stockholmresilience.su.se}, Susann Müller \email{susann.müller@ufz.de}
}
\references{\emph{Christin Koch, Ingo Fetzer, Hauke Harms, and Susann Müller. CHIC - An Automated
Approach for the Detection of Dynamic Variations in Complex Microbial Communities.
Cytometry Part A, 2013.}}
%\keyword{ package 

%}
%\seealso{

%}
%\examples{

%}