\name{Results_overlaps}
\docType{data}
\encoding{UTF-8}
\alias{Results_overlaps}
\title{Dataset with Results_overlaps}
\description{Example dataset of Results_overlaps calculated by ImageJ (see reference for further information).}
\usage{data(Results_overlaps)}
\format{Data frame with 190 observations of 2 variables.}
\references{\emph{Christin Koch, Ingo Fetzer, Hauke Harms and Susann Müller. CHIC - an automated approach for the detection of dynamic variations in complex microbial communities. Cytometry Part A, 2013.}
}
\keyword{datasets}