\name{Results_xor}
\docType{data}
\encoding{UTF-8}
\alias{Results_xor}
\title{Dataset with Results_xor}
\description{Example dataset of Results_xor calculated by ImageJ (see reference for further information).}
\usage{data(Results_xor)}
\format{Data frame with 190 observations of 3 variables.}
\references{\emph{Christin Koch, Ingo Fetzer, Hauke Harms and Susann Müller. CHIC - an automated approach for the detection of dynamic variations in complex microbial communities. Cytometry Part A, 2013.}
}
\keyword{datasets}